add dto

+++ onlineshop/web/src/main/java/by/masalsky/onlineshop/controllers/UserController.java
@@ -1,7 +1,6 @@
 package by.masalsky.onlineshop.controllers;
 
 import by.masalsky.onlineshop.dto.UserDto;
-import by.masalsky.onlineshop.entities.User;
 import by.masalsky.onlineshop.services.interfaces.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -20,39 +19,33 @@ public class UserController {
     @Autowired
     private IUserService userService;
 
-    @RequestMapping(value = "/q")
-    public String home(){
-        return "Hello dimaQ!";
-    }
-
-    @RequestMapping(value = "/qw")
-    public String homeW(){
-        return "Hello dimaW!";
-    }
-
     @RequestMapping(value = "/", method = RequestMethod.GET)
-    public ResponseEntity<List<User>> getAllUsers() {
-        List<User> list = userService.getAll();
-        return new ResponseEntity<List<User>>(list, HttpStatus.OK);
+    public ResponseEntity<List<UserDto>> getAllUsers() {
+        List<UserDto> list = userService.getAll();
+        return new ResponseEntity<>(list, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public ResponseEntity<User> getArticleById(@PathVariable("id") Integer id) {
-        User user = userService.getById(id);
-        return new ResponseEntity<User>(user, HttpStatus.OK);
+    public ResponseEntity<UserDto> getUserById(@PathVariable("id") Integer id) {
+        UserDto user = userService.getById(id);
+        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Void> deleteUser(@PathVariable("id") Integer id) {
-        userService.delete(id);
-        return new ResponseEntity<Void>(HttpStatus.OK);
+        UserDto user = userService.getById(id);
+        if (user != null) {
+            userService.delete(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
     }
 
-    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    @RequestMapping(value = "/", method = RequestMethod.POST)
     public ResponseEntity createUser(@RequestBody UserDto userDto) {
-        User user = userService.getByLogin(userDto.getLogin());
+        UserDto user = userService.getByLogin(userDto.getLogin());
         if (user == null) {
-            userService.registration(userDto);
+            userService.save(userDto);
             return new ResponseEntity<Void>(HttpStatus.OK);
         } else
             return new ResponseEntity<Void>(HttpStatus.CONFLICT);
@@ -61,12 +54,13 @@ public class UserController {
     @RequestMapping(value = "/", method = RequestMethod.PUT)
     public ResponseEntity updateUser(@RequestBody UserDto userDto) {
         userService.update(userDto);
-        return new ResponseEntity<UserDto>(userDto, HttpStatus.OK);
+        userDto = userService.getById(userDto.getId());
+        return new ResponseEntity<>(userDto, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/login", method = RequestMethod.POST)
     public ResponseEntity loginUser(@RequestBody UserDto userDto) {
         boolean flag = userService.isAuthorized(userDto.getLogin(), userDto.getPassword());
-        return new ResponseEntity<Boolean>(flag, HttpStatus.OK);
+        return new ResponseEntity<>(flag, HttpStatus.OK);
     }
 }
